Move `MonkeySpec` to its own class.
