+++ plugin/src/main/groovy/com/novoda/gradle/command/Monkey.groovy
@@ -13,9 +13,7 @@ class Monkey extends AdbTask {
 
     @TaskAction
     void exec() {
-        Spec monkey = pluginEx.monkey
-
-        logger.info monkey.categories.toString()
+        MonkeySpec monkey = pluginEx.monkey
 
         def arguments = ['shell', 'monkey']
         arguments += ['-p', packageName]
@@ -25,41 +23,5 @@ class Monkey extends AdbTask {
             arguments += ['-s', monkey.seed]
         }
         assertDeviceAndRunCommand(arguments)
-    }
-
-    static class Spec {
-
-        private static final int EVENTS_DEFAULT = 10000
-
-        def events
-        def seed
-        def categories = []
-
-        void events(events) {
-            this.events = events
-        }
-
-        void seed(seed) {
-            this.seed = seed
-        }
-
-        void categories(... categories) {
-            this.categories.addAll(categories)
-        }
-
-        // prefer system property over direct setting to enable commandline arguments
-        def getEvents() {
-            System.properties['events'] ?: events ?: EVENTS_DEFAULT
-        }
-
-        // prefer system property over direct setting to enable commandline arguments
-        def getCategories() {
-            def systemCategories = System.properties['categories']
-            systemCategories ? [systemCategories] : categories
-        }
-
-        def getSeed() {
-            System.properties['seed'] ?: seed
-        }
     }
 }
